Rework datasource into snapshot package.
Add Snapshot interface.

Create snapshot struct which constains a map of slices instead of
listers.
