+++ pkg/analyze/snapshot/archivereader.go
@@ -1,18 +1,12 @@
-package analyze
+package snapshot
 
 import (
-	"context"
 	"fmt"
 	"io/fs"
-	"os"
 	"path/filepath"
 	"reflect"
 
-	scyllav1 "github.com/scylladb/scylla-operator/pkg/api/scylla/v1"
-	scyllav1listers "github.com/scylladb/scylla-operator/pkg/client/scylla/listers/scylla/v1"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	corev1listers "k8s.io/client-go/listers/core/v1"
 	"k8s.io/client-go/tools/cache"
 )
 
@@ -24,6 +18,38 @@ func getIndexerForType(indexers map[reflect.Type]cache.Indexer, objType reflect.
 	indexer := cache.NewIndexer(cache.MetaNamespaceKeyFunc, cache.Indexers{})
 	indexers[objType] = indexer
 	return indexer
+}
+
+func NewSnapshotFromFS(fsys fs.FS, decoder runtime.Decoder) (Snapshot, error) {
+	ds := NewEmptySnapshot()
+
+	err := fs.WalkDir(fsys, ".", func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		if d.IsDir() || filepath.Ext(path) != ".yaml" {
+			return nil
+		}
+		content, err := fs.ReadFile(fsys, path)
+		if err != nil {
+			return fmt.Errorf("can't read file %q: %w", path, err)
+		}
+		obj, _, err := decoder.Decode(content, nil, nil)
+
+		if err != nil {
+			if !runtime.IsNotRegisteredError(err) {
+				return fmt.Errorf("can't deserialize file %q: %w", path, err)
+			}
+			return nil
+		}
+		ds.Add(obj)
+		return nil
+	})
+
+	if err != nil {
+		return nil, fmt.Errorf("can't walk the file tree: %w", err)
+	}
+	return &ds, nil
 }
 
 func IndexersFromFS(fsys fs.FS, decoder runtime.Decoder) (map[reflect.Type]cache.Indexer, error) {
@@ -61,20 +87,4 @@ func IndexersFromFS(fsys fs.FS, decoder runtime.Decoder) (map[reflect.Type]cache
 		return nil, fmt.Errorf("can't walk the file tree: %w", err)
 	}
 	return indexers, nil
-}
-
-func NewDataSourceFromFS(ctx context.Context, archivePath string, decoder runtime.Decoder) (*DataSource, error) {
-	indexers, err := IndexersFromFS(os.DirFS(archivePath), decoder)
-	if err != nil {
-		return nil, fmt.Errorf("can't build indexers from fs: %w", err)
-	}
-
-	return &DataSource{
-		PodLister:            corev1listers.NewPodLister(getIndexerForType(indexers, reflect.TypeOf(&corev1.Pod{}))),
-		ServiceLister:        corev1listers.NewServiceLister(getIndexerForType(indexers, reflect.TypeOf(&corev1.Service{}))),
-		SecretLister:         corev1listers.NewSecretLister(getIndexerForType(indexers, reflect.TypeOf(&corev1.Secret{}))),
-		ConfigMapLister:      corev1listers.NewConfigMapLister(getIndexerForType(indexers, reflect.TypeOf(&corev1.ConfigMap{}))),
-		ServiceAccountLister: corev1listers.NewServiceAccountLister(getIndexerForType(indexers, reflect.TypeOf(&corev1.ServiceAccount{}))),
-		ScyllaClusterLister:  scyllav1listers.NewScyllaClusterLister(getIndexerForType(indexers, reflect.TypeOf(&scyllav1.ScyllaCluster{}))),
-	}, nil
 }
